Fix formating issue
printf(2) expects precision argument of 'int' type on the other hand
pointer arithmetic could produce a 'long int' result on some platforms,
what finally leads to the compilation warnings. Fix this by explicitly
converting pointer arithmetic result to 'int' before passing it to
printf(2) call.

+++ unnpk.c
@@ -219,7 +219,7 @@ static int proc_part_data_pkg_desc(uint8_t *data, const uint32_t size,
 	while (*e == '\n' || *e == ' ')
 		e--;
 
-	printf("Description: %.*s\n", e - p + 1, p);
+	printf("Description: %.*s\n", (int)(e - p + 1), p);
 
 	return 0;
 }
